backend/requirements.txt
fastapi[all]==0.99.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.45
alembic==1.11.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
redis==5.3.1
aioredis==2.0.1
